+++ tests/src/rules/dynamic-import-chunkname.js
@@ -5,7 +5,7 @@ import semver from 'semver';
 const rule = require('rules/dynamic-import-chunkname');
 const ruleTester = new RuleTester();
 
-const commentFormat = '[0-9a-zA-Z-_/.]+';
+const commentFormat = '([0-9a-zA-Z-_/.]|\\[(request|index)\\])+';
 const pickyCommentFormat = '[a-zA-Z-_/.]+';
 const options = [{ importFunctions: ['dynamicImport'] }];
 const pickyCommentOptions = [{
@@ -21,8 +21,9 @@ const noLeadingCommentError = 'dynamic imports require a leading comment with th
 const nonBlockCommentError = 'dynamic imports require a /* foo */ style comment, not a // foo comment';
 const noPaddingCommentError = 'dynamic imports require a block comment padded with spaces - /* foo */';
 const invalidSyntaxCommentError = 'dynamic imports require a "webpack" comment with valid syntax';
-const commentFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: ["']${commentFormat}["'],? */`;
-const pickyCommentFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: ["']${pickyCommentFormat}["'],? */`;
+const commentFormatError = `dynamic imports require a "webpack" comment with valid syntax`;
+const chunkNameFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: ["']${commentFormat}["'],? */`;
+const pickyChunkNameFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: ["']${pickyCommentFormat}["'],? */`;
 
 ruleTester.run('dynamic-import-chunkname', rule, {
   valid: [
@@ -53,6 +54,34 @@ ruleTester.run('dynamic-import-chunkname', rule, {
         'someModule'
       )`,
       options: pickyCommentOptions,
+    },
+    {
+      code: `dynamicImport(
+        /* webpackChunkName: "[request]" */
+        'someModule'
+      )`,
+      options,
+    },
+    {
+      code: `dynamicImport(
+        /* webpackChunkName: "my-chunk-[request]-custom" */
+        'someModule'
+      )`,
+      options,
+    },
+    {
+      code: `dynamicImport(
+        /* webpackChunkName: '[index]' */
+        'someModule'
+      )`,
+      options,
+    },
+    {
+      code: `dynamicImport(
+        /* webpackChunkName: 'my-chunk.[index].with-index' */
+        'someModule'
+      )`,
+      options,
     },
     {
       code: `import(
@@ -130,6 +159,24 @@ ruleTester.run('dynamic-import-chunkname', rule, {
     },
     {
       code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackPrefetch: 12 */
+        'test'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackPrefetch: -30 */
+        'test'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
         /* webpackChunkName: 'someModule' */
         'someModule'
       )`,
@@ -142,6 +189,217 @@ ruleTester.run('dynamic-import-chunkname', rule, {
         'someModule'
       )`,
       options: pickyCommentOptions,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "[request]" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "my-chunk-[request]-custom" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: '[index]' */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: 'my-chunk.[index].with-index' */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackInclude: /\\.json$/ */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule", webpackInclude: /\\.json$/ */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackExclude: /\\.json$/ */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule", webpackExclude: /\\.json$/ */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackPreload: true */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackPreload: 0 */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackPreload: -2 */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule", webpackPreload: false */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackIgnore: false */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule", webpackIgnore: true */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackMode: "lazy" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: 'someModule', webpackMode: 'lazy' */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackMode: "lazy-once" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackMode: "eager" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackMode: "weak" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackExports: "default" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule", webpackExports: "named" */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackExports: ["default", "named"] */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: 'someModule', webpackExports: ['default', 'named'] */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "someModule" */
+        /* webpackInclude: /\\.json$/ */
+        /* webpackExclude: /\\.json$/ */
+        /* webpackPrefetch: true */
+        /* webpackPreload: true */
+        /* webpackIgnore: false */
+        /* webpackMode: "eager" */
+        /* webpackExports: ["default", "named"] */
+        'someModule'
+      )`,
+      options,
       parser,
     },
     ...SYNTAX_CASES,
@@ -256,6 +514,54 @@ ruleTester.run('dynamic-import-chunkname', rule, {
     },
     {
       code: `import(
+        /* webpackChunkName: true */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackChunkName: true */
+        'someModule'
+      )`,
+      errors: [{
+        message: chunkNameFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackChunkName: "my-module-[id]" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackChunkName: "my-module-[id]" */
+        'someModule'
+      )`,
+      errors: [{
+        message: chunkNameFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackChunkName: ["request"] */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackChunkName: ["request"] */
+        'someModule'
+      )`,
+      errors: [{
+        message: chunkNameFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
         /*webpackChunkName: "someModule"*/
         'someModule'
       )`,
@@ -364,7 +670,183 @@ ruleTester.run('dynamic-import-chunkname', rule, {
         'someModule'
       )`,
       errors: [{
-        message: pickyCommentFormatError,
+        message: pickyChunkNameFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackPrefetch: "module", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackPrefetch: "module", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackPreload: "module", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackPreload: "module", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackIgnore: "no", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackIgnore: "no", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackInclude: "someModule", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackInclude: "someModule", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackInclude: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackInclude: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackExclude: "someModule", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackExclude: "someModule", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackExclude: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackExclude: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackMode: "fast", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackMode: "fast", webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackMode: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackMode: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackExports: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackExports: true, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackExports: /default/, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackExports: /default/, webpackChunkName: "someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: commentFormatError,
         type: 'CallExpression',
       }],
     },
@@ -478,7 +960,7 @@ ruleTester.run('dynamic-import-chunkname', rule, {
         'someModule'
       )`,
       errors: [{
-        message: pickyCommentFormatError,
+        message: pickyChunkNameFormatError,
         type: 'CallExpression',
       }],
     },
@@ -570,19 +1052,226 @@ context('TypeScript', () => {
         {
           code: `import(
             /* webpackChunkName: "someModule" */
+            /* webpackPrefetch: 11 */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackPrefetch: -11 */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
             'someModule'
           )`,
           options: pickyCommentOptions,
           parser: typescriptParser,
-          errors: [{
-            message: pickyCommentFormatError,
-            type: nodeType,
-          }],
         },
         {
           code: `import(
             /* webpackChunkName: 'someModule' */
             'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "[request]" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "my-chunk-[request]-custom" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: '[index]' */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: 'my-chunk.[index].with-index' */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackInclude: /\\.json$/ */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule", webpackInclude: /\\.json$/ */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackExclude: /\\.json$/ */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule", webpackExclude: /\\.json$/ */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackPreload: true */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule", webpackPreload: false */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackIgnore: false */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule", webpackIgnore: true */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackMode: "lazy" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: 'someModule', webpackMode: 'lazy' */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackMode: "lazy-once" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackMode: "eager" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackMode: "weak" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackExports: "default" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule", webpackExports: "named" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackExports: ["default", "named"] */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: 'someModule', webpackExports: ['default', 'named'] */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackInclude: /\\.json$/ */
+            /* webpackExclude: /\\.json$/ */
+            /* webpackPrefetch: true */
+            /* webpackPreload: true */
+            /* webpackIgnore: false */
+            /* webpackMode: "eager" */
+            /* webpackExports: ["default", "named"] */
+            'someModule'
           )`,
           options,
           parser: typescriptParser,
@@ -795,6 +1484,54 @@ context('TypeScript', () => {
         },
         {
           code: `import(
+            /* webpackChunkName: true */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName: true */
+            'someModule'
+          )`,
+          errors: [{
+            message: chunkNameFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "my-module-[id]" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName: "my-module-[id]" */
+            'someModule'
+          )`,
+          errors: [{
+            message: chunkNameFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName: ["request"] */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName: ["request"] */
+            'someModule'
+          )`,
+          errors: [{
+            message: chunkNameFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
             /* webpackChunkName: "someModule123" */
             'someModule'
           )`,
@@ -805,7 +1542,183 @@ context('TypeScript', () => {
             'someModule'
           )`,
           errors: [{
-            message: pickyCommentFormatError,
+            message: pickyChunkNameFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackPrefetch: "module", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackPrefetch: "module", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackPreload: "module", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackPreload: "module", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackIgnore: "no", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackIgnore: "no", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackInclude: "someModule", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackInclude: "someModule", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackInclude: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackInclude: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackExclude: "someModule", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackExclude: "someModule", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackExclude: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackExclude: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackMode: "fast", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackMode: "fast", webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackMode: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackMode: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackExports: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackExports: true, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackExports: /default/, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackExports: /default/, webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
             type: nodeType,
           }],
         },